refactor: share api client between dashboard & settings-form
