+++ dashboard/src/shared
@@ -1,0 +1,1 @@
+../../shared/src
