+++ shared/README.md
@@ -1,0 +1,3 @@
+# shared
+
+This library was generated with [Nx](https://nx.dev).
