+++ shared/project.json
@@ -1,0 +1,16 @@
+{
+  "name": "shared",
+  "$schema": "../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "shared/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["shared/**/*.ts"]
+      }
+    }
+  },
+  "tags": []
+}
